requirements.txt
streamlit==1.38.0
pdf2image==1.17.0
Pillow==10.4.0
openai==1.51.0
python-dotenv==1.0.1
pandas==2.1.4
psutil==5.9.6
pathlib2==2.3.7